+++ src/build.ts
@@ -1,121 +1,87 @@
-import { build as esbuild, BuildOptions, Plugin } from 'esbuild'
-import fs from 'fs'
-import path from 'path'
-import { argv2config, lookupFile, resolveExternal } from './utils'
+import esbuild, { BuildOptions } from "esbuild";
+import path from "path";
+import util from "util";
+import error from "./error.txt";
+import { lookupExternal, lookupFile, tmpdir } from "./utils";
 
-interface Options {
-  plugins: Plugin[]
-  entryPoints: string[]
-  args: string[]
+type NodeVersion = "14" | "15" | "16";
+type Target = `node${NodeVersion}`;
+
+const nodeVersion = process.versions.node.split(".", 2)[0] as NodeVersion;
+const target = `node${nodeVersion}` as Target;
+
+export type Format = "esm" | "cjs";
+const extname: Record<Format, ".mjs" | ".js"> = { esm: ".mjs", cjs: ".js" };
+
+export async function build(entry: string, format: Format = "esm", options?: BuildOptions) {
+  const dirname = path.dirname(entry);
+  const pkgPath = lookupFile("package.json", dirname);
+  const outfile = path.join(tmpdir(), entry + extname[format]);
+  const result = await esbuild.build({
+    entryPoints: [entry],
+    external: lookupExternal(pkgPath),
+    platform: "node",
+    target,
+    format,
+    bundle: true,
+    sourcemap: true,
+    outfile,
+    ...options,
+  });
+  return { outfile, result };
 }
 
-const defaultOptions: BuildOptions = {
-  platform: 'node',
-  target: 'node12',
-  bundle: true,
-  minify: true,
-  sourcemap: true,
-  outdir: 'dist',
-  logLevel: 'info',
+export function errorMessage(file: string, args: string[]) {
+  const template = { file, args: args.map(e => util.inspect(e)).join(" ") };
+  return error.replace(/{(\w+)}/g, (_, key: "file" | "args") => template[key] || "");
 }
 
-export async function build({ plugins, entryPoints, args }: Options) {
-  const options: BuildOptions = {
-    ...defaultOptions,
-    entryPoints,
-    external: resolveExternal(false),
-    plugins,
-    ...argv2config(args),
-  }
-
-  if (entryPoints.length === 0) {
-    return await rollpkg(options)
-  }
-
-  fixConflicts(options)
-
-  // Can only support add this banner in one entry point
-  if (entryPoints.length === 1 && entryPoints[0].includes('bin')) {
-    setBinBanner(options)
-  }
-
-  return await esbuild(options)
-}
-
-function setBinBanner(options: BuildOptions) {
-  options.banner = { js: '#!/usr/bin/env node' }
-}
-
-function fixConflicts(options: BuildOptions) {
-  if (options.outfile && options.outdir) {
-    // Cannot use both "outfile" and "outdir"
-    delete options.outdir
-  }
-}
-
-function rollpkg(options: BuildOptions) {
-  const pkgPath = lookupFile(process.cwd(), 'package.json')
-  if (!pkgPath) {
-    console.error('can not use --build <no entry> outside of npm package')
-    process.exit(1)
-  }
-
-  const pkgDir = path.dirname(pkgPath)
-  const pkg = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'))
-
-  const first = (...names: string[]) => {
-    const exts = 'tsx|ts|jsx|mjs|js'.split('|')
-    for (const name of names) {
-      for (const ext of exts) {
-        const file = path.join(pkgDir, `${name}.${ext}`)
-        if (fs.existsSync(file) && fs.statSync(file).isFile()) {
-          return path.relative(process.cwd(), file)
+const cliRE = /^--([-a-z]+)(?::([^=]+))?(?:=(.+))?$/;
+const booleans = { true: true, false: false } as const;
+export function argsToBuildOptions(args: string[]) {
+  const buildOptions: BuildOptions = {};
+  for (const arg of args) {
+    const m = cliRE.exec(arg);
+    if (m) {
+      const [, slashKey, subKey, value] = m;
+      const key = camelize(slashKey) as keyof BuildOptions;
+      if (subKey && value) {
+        // define, loader, banner, footer, stdin
+        // their values are all string
+        (buildOptions[key] ||= {} as any)[subKey] = value;
+      } else if (subKey) {
+        // pure, external, inject
+        // their values are all string
+        (buildOptions[key] ||= [] as any).push(subKey);
+      } else if (value) {
+        // resolveExtensions, mainFields, conditions, target
+        // their values can have ',' to become array
+        // others are boolean | string
+        const bool: boolean | undefined = booleans[value as keyof typeof booleans];
+        if (bool !== undefined) {
+          // --a=false
+          buildOptions[key] = bool as any;
+        } else if (value.includes(",")) {
+          // --a=b,c
+          buildOptions[key] = value.split(",") as any;
+        } else {
+          // --a=b
+          buildOptions[key] = value as any;
         }
+      } else {
+        buildOptions[key] = true as any;
       }
+    } else {
+      throw new Error(`unknown arg: ${arg}`);
     }
   }
-  const format = (key: string) =>
-    pkg.type === 'module' ||
-    key.endsWith('.mjs') ||
-    key.endsWith('.es.js') ||
-    key.endsWith('.esm.js')
-      ? 'esm'
-      : 'cjs'
+  return buildOptions;
+}
 
-  const inOut: { in: string; out: string; format: 'esm' | 'cjs' }[] = []
+function camelize(s: string) {
+  return s.split("-").reduce((a, b) => a + capitalize(b));
+}
 
-  let file: string | undefined
-  if (pkg.main && (file = first('src/index', 'index'))) {
-    inOut.push({ in: file, out: pkg.main, format: format(pkg.main) })
-  }
-  if (pkg.module && file) {
-    inOut.push({ in: file, out: pkg.module, format: format(pkg.module) })
-  }
-  if (pkg.bin && (file = first('src/bin', 'bin'))) {
-    let out = pkg.bin
-    if (Array.isArray(out)) {
-      out = out[0]
-    } else if (typeof out === 'object' && out !== null) {
-      out = Object.values(out)[0]
-    }
-    if (typeof out === 'string') {
-      inOut.push({ in: file, out, format: format(out) })
-    }
-  }
-
-  return Promise.all(
-    inOut.map(e => {
-      const buildOptions = {
-        ...options,
-        entryPoints: [e.in],
-        outfile: e.out,
-        format: e.format,
-      }
-      fixConflicts(buildOptions)
-      if (e.in.includes('bin')) {
-        setBinBanner(buildOptions)
-      }
-      return esbuild(buildOptions)
-    })
-  )
+function capitalize(s: string) {
+  return s.charAt(0).toUpperCase() + s.slice(1);
 }