+++ src/shim.d.ts
@@ -1,4 +1,3 @@
 declare module "*.txt" {
-  const content: string;
-  export default content;
+  export default "";
 }